Automated g4 rollback of changelist 85646980.
*** Reason for rollback ***

Need to roll it back. Will be back with it soon. There's one situation where this breaks when click arrives on a different target vs touchend.

*** Original change description ***

Change fast click implementation to be more transparent with respect to the natural click event that is emitted by the browser.
This event should be ignored by the application, but still get a chance to executed its default action if the application did not call prevent default on the synthetic click event.

***

	Change on 2015/02/05 by dvoytenko <dvoytenko@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=85671226