Change fast click implementation to be more transparent with respect to the natural click event that is emitted by the browser.
This event should be ignored by the application, but still get a chance to execute its default action if the application did not call prevent default on the synthetic click event.

The difference with the previous CL that was rolled back is the explicit handling of a situation when click's target is different from touchend. This is the real situation that flows from some inconsistencies in browser specs. In particular both Android Chrome and iOS Safari are affected. Jsbin (http://jsbin.com/zeriwa/2/) shows the inconsistencies in targets. You can run it on Android Chrome - it will print target differences in the page.

	Change on 2015/02/09 by dvoytenko <dvoytenko@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=85938539