Fast click: alternative implementation. This solution relies on issuing a new click event immediately rather than waiting for the native click to arrive. This solution is superior to the previous one in most of respects: it's deterministic, non-flaky and much simpler. The only drawback of this solution (and why we tried the other approach first) was b/c it "touchend" by default handles blurring of the active element better and as the result closes virtual keyboard.
However, we found out that it's better to emulate this behavior than to deal with inconsistencies of capturing of mouse events. In particular we realized that due to mobile browser mouse event emulation, we would also need to block mousedown, mousemove and mouseover in the old scheme).

	Change on 2015/02/10 by dvoytenko <dvoytenko@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=86045351