Adding mouse position support to native mousedown, mouseup and click events. This is done analog to javascript/closure/testing/events/events.js that seems to have been a template for nativeevents.js.
This is a roll forward of [] The culprit was a test that expected (0, 0) coordinates if no explicit point is provided, but the new version of fireMouseDownEvent() uses the target position as default. This is resolved by providing explicit (0, 0) coordinates.

	Change on 2015/07/08 by sschuster <sschuster@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=97765998