Update the ActionFlow to fire a CREATED event on the jsaction.ActionFlow.report object whenever an action flow is instantiated. Listeners receive an instance of the newly created flow.
This CL is necessary to allow applications to build a single point of entry for
dealing with new flows. [] will use this event to allow a FlowGate component
to deal with families of flows in a specific way. For more information, see:
https://goto.google.com/[]-gated-flows.

	Change on 2015/08/11 by joshharrison <joshharrison@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=100421361