Migrate from --output_wrapper='(function(){%output%})();' to --isolation_mode=IIFE.
The two are equivalent except that the new version works in []
(equivalent to  --output_wrapper='(function(){%output%}).call(this);').
Specifically, in strict mode the global "this" is not implicitly available
within function, we explicitly bind the "this" to preserve the current behavior
and to ease the adoption of strict mode output from the compiler.

[]

Tested:
    TAP --sample for global presubmit queue
    []
	Change on 2016/12/08 by blickly <blickly@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=141526621