Clicks should focus elements in jsaction.
When fastclick recreates a touch event as a click([] it doesn't end up focusing the element which is clicked, this leaves you with a weird case where certain things which should be focused are not.

Example:

Broken,
https://screenshot.[].com/zTOukiV7UJ8
https://www.google.com/search?q=flowers&expflags=Jsaction__fastclick_enable:true (use chrome emulator also to get tap events)
Click on WTA icon and see 'document.activeElement' evaluates to <body/>

Working
https://screenshot.[].com/zTOukiV7UJ8
https://www.google.com/search?q=flowers&expflags=Jsaction__fastclick_enable:false (use chrome emulator also to get tap events)
Click on WTA icon and see 'document.activeElement' evaluates to the element with the jsaction.

I'm fairly confident that this is the correct place to put this code as we are trying emulate a click which should be focusing focusable elements.

Demo w/ fix:
[]

See cl/197509579 for more info.

RELNOTES: Fixes bug where simulated clicks with fastclick implementation were not focusing the element.

	Change on 2018/06/22 by jackgresset <jackgresset@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=201698162