Fix fastclick preventing blurring of activeElement
Fastclick preventDefaults the touchend event so that it cancels the browser 'click' event which would be a duplicate of its synthetic click event. But this also prevents the browser from unfocusing the active element when the user clicks elsewhere on the page.

RELNOTES: Fix fastclick preventing blurring of activeElement

	Change on 2019/08/05 by rahatahmed <rahatahmed@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=261791386