Setup event contract and dispatcher to handle a11y-click-casting in the dispatcher.
These changes add a new compile flag to event contract and mod to dispatcher that defer parsing key-events to the dispatcher. The keypress-parsing is required for treating space/enter presses as clicks for A11Y_CLICK_SUPPORT, but add ~1.6KB to the header in []. We will use the compile flag to build a binary without this logic and run an experiment to see what effects is has on bytes-sent and latency.

The goal for these changes is to keep behavior consistent for existing (internal and external) clients of jsaction. Even with the changes enabled, users should see the same affects of having the A11Y_CLICK_SUPPORT flag enabled.

Demos
-----
Bring keyboard focus expandos and trigger with 'enter'. Admittedly this may not be the best feature to demo against, but I found it to work relatively well for mobile.
-----
With flags enabled: []
Control: []
-----

Background: []
Design doc: []

TGP: []
(Failures unrelated)

RELNOTES: n/a

	Change on 2019/10/03 by sulawko <sulawko@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=272741072