Automated g4 rollback of changelist 271615353.
*** Reason for rollback ***

[] JSDT metric has been at 100% since this change. Rolling back to investigate.

*** Original change description ***

Change the default return value of ActionFlow.getDelay to -1 (rather than 0) when the event or its originalTimestamp property is null or undefined. Since 0ms is a valid delay, we do not want to confuse valid and invalid responses of 0ms. We want to determine whether or not we are over-reporting the number of first-input delays with a value of 0ms and this change will ensure that we are only reporting valid 0ms delays.

RELNOTES: n/a

***

#frameworks-web-onduty

	Change on 2019/10/18 by csmick <csmick@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=275577510